platform: tegra: mc: fix coverity defects
1. specify the maximum characters for sscanf/strncpy to
prevent buffer overflow issues.
2. use explicit typecasts to prevent overflowing expression

Coverity ID 31114 30994 24124

bug 200192541

Reviewed-on: http://git-master/r/1130363
(cherry picked from commit 91dacbb833225e462e9aecc90fb2060d04f408f1)
Reviewed-on: http://git-master/r/1140172
(cherry picked from commit 7158b4df6d69cb71f7c3e30ffeda4363afba884e)
Reviewed-on: http://git-master/r/1150603

+++ drivers/platform/tegra/mc/tegra21x_la.c
@@ -642,7 +642,7 @@ static int t21x_handle_disp_la(enum tegra_la_id id,
 	if (effective_mccif_buf_sz < 0)
 		return -1;
 
-	la_bw_upper_bound_nsec_fp = effective_mccif_buf_sz *
+	la_bw_upper_bound_nsec_fp = (long long)effective_mccif_buf_sz *
 					LA_FP_FACTOR /
 					bw_mbps;
 	la_bw_upper_bound_nsec_fp = la_bw_upper_bound_nsec_fp *
